refactor from thermo-nuclear-code-review, add tests, comments
